refactor: move app.tsx to src/
