+++ src/index.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 
-import { App } from './foundation/app';
+import { App } from './app';
 
 ReactDOM.render(<App />, document.getElementById('app'));
